Unicode rethink: Uglify ==; big speedups.
Begin uglifying String's ==. Every same-representation case has
opportunities for speedup (and, actually, so does many
different-representations).

This yields us a nice 50x speedup for == when comparing small strings
with the same small representation and the same bitsPerComponent. It
also will serve as a basis for more special-case checking.

It is very ugly, and will be getting uglier, but once we meet perf
goals I'll look into prettying it up without sacrificing perf.