[Constraint solver] Tweak test based on using a bit more memory for -propagate-constraints.
This test is trying to confirm that memory usage is independent and that
follow-on expressions do not fail just because of the first failure, and
now we naturally fail on another expression because of the additional
memory used for -propagate-constraints.

Tweak the test to bump up the threshold and swap the order of
expressions so that the now-failing expression is first.