IRGen: Allow pure-ObjC refcounting class existentials
Class-constrained existentials usually have unknown
reference counting in Swift, because they can contain
either Objective-C or native Swift classes.

But if the class existential has a superclass bound
which is known to be an imported Objective-C class,
we can use Objective-C reference counting instead of
unknown reference counting.

This is tested with the next commit that adds
ClangImporter support.