Mangle imported declarations using their C names.
This makes them consistent no matter what shenanigans are pulled by
the importer, particularly NS_ENUM vs. NS_OPTIONS and NS_SWIFT_NAME.

The 'NSErrorDomain' API note /nearly/ works with this, but the
synthesized error struct is still mangled as a Swift declaration,
which means it's not rename-stable. Fortunately, these are rare to
appear at API boundaries.

The one place where this still falls down is NS_STRING_ENUM: when this
is applied, a typedef is imported as a unique struct, but without it
it's just a typealias for the underlying type. I don't have a solution
for this yet.

rdar://problem/31616162