[Constraint solver] Enable constraint propagation pass.
Enable the constraint propagation pass in an effort to do a pre-pass to
reduce the amount of work we need to do during constraint solving.

There are still improvements that can be made here, but this is
currently passing tests and showing benefit in cases where we have
overloads taking concrete types, so it's worth enabling and then
incrementally improving.

Despite dramatically speeding up some test cases, because this is a
prepass it is going to result in a marginal slow-down in type checking
of some code. The type checking time for the standard library is
approximately 14% slower on my machine with this enabled (14.06s
vs. 16.03s).