[Stdlib] Rationalize the SubSequence type of a filtered Collection.
Address ABI FIXMEs #28, #80, #81 by making the SubSequence of a filtered
collection into be a filtered collection of the base's
SubSequence rather than a slice of this filtered collection, e.g.,
replace `Slice<LazyFilterCollection<Array<T>>>` with
`LazyFilterCollection<ArraySlice<T>>`, which allows slicing
optimizations of the base type to kick in.

While here, eliminate the vestigial type `LazyFilterIndex`, which was
necessary pre-Swift-3 to allow the index to move. Swift 3's indexing
model means that the movement of indices is on the collection itself,
so we no longer need `LazyFilterIndex`: instead, the `Index` type of
the lazy filtered collection is simply the `Index` type of the base
collection, which is a nice convenience: it means you can take indices
from a lazy wrapper around a given collection C and use them with the
collection C (and, with care, vice-versa) without jumping through
extra hoops.