[Constraint solver] While simplifying constraints, test protocol conformance of arguments.
When we simplify bindings and function applications, we generate
independent constraints for protocol conformance and argument
conversion. What we previously did not do is examine these constraints
together to see if they make any sense.

This commit collects the protocol conformance constraints while
simplifying these independent constraints, and then tests these against
the argument conversion constraints for arguments to see if the
arguments could possibly conform to the protocols.

The end result is eliminating some of the exponential behavior observed
the type checker.

There is a minor unexpected regression in some diagnostics around
operations mixing String and NSString. I have updated the tests to
the still correct but less specific diagnostic for the moment and will
open a bug to investigate and fix these.