[TSan/Exclusivity] Make test more robust against optimization
Add a "black hole" uninstrumented function to the test that prevents the
compiler from removing an unused read that needs to be instrumented by
TSan for the test to pass.

This is needed to prevent the optimizer from removing a useless read
when static enforcement of exclusivity is turned on.

(cherry picked from commit 112c3c65c0505e874e5650e436dd981b7cd418e8)