Prototype stateful Unicode decoding
Having looked at the old transcoding code, it's very clear that no stateless
scheme has a hope of competing with it on performance.  We just need reverse
transcoding and a way to hook up indices (I believe we can use IndexingIterator
for this, that's coming).

This code demonstrates UTF-8 (the hard one) and is competitive on performance
with the existing code.  I believe it could be even better-performing if the
cases that do parsing were connected directly with the cases that do decoding.

Finally, tatoo this on your forehead: Dmitri Shall Not Be Underestimated