IRGen: Fix type conversion of protocol compositions containing a generic superclass constraint
Do the same thing we do for nominal types, where if the type
contains an archetype, just mangle the unbound generic form,
since the name is just for documentation purposes in LLVM IR
dumps and does not have to be unique.

This is tested as part of an upcoming patch.