Fix the tbi test even harder.
A few things here:

1. I changed the .swift file to be a .sil file. This one is for Jordan. It
should also make the test not run into any issues due to swift level codegen
changes. So we should have an even more robust test.

2. I fixed the watchossimulator test by eliminating any dependency of this test
on an SDK. This is a pure compiler/target test. Just like the stdlib is not
necessary, SDKs are not necessary either.